backend/requirements.txt
Django>=6.0
djangorestframework
django-cors-headers
psycopg2-binary
